can specify fill_byte for function write_to_memory
If the string given to write_to_memory is not a multiple of 4 (or 8?), it will need to be padded by a fill byte. The fill byte was set to 0xff. This commit makes it possible to give the fill byte as an argument.

E.g. useful if we want to write bash strings, and fill it with spaces or hashtags.