Remove constrained registers from popped_regs set
If a gadget pops a value into a register but the value is used in a
branch condition or memory access address, we can't fully control it so
we remove it from the set of popped registers. This also makes
gadget.constraint_regs include registers that affect memory access
addresses.