Update tests to check chain.next_pc_idx()
If chain.next_pc_idx() is not None, the address to jump to after the
chain should be placed at that index instead of after the chain. In
cases like this the chain concretization function will attempt to append
a no-op gadget so that the chain jumps to the address immediately after
it, but this is not always possible on architectures without a stack
return instruction.